refactored MonitorReport components
