+++ platform/src/common/components/Collocation/Report/MonitorReport/DataCompletenessTable/DataTable.jsx
